Update data.js
I updated it so we can easily access a bye week or not. I also added to see whether they were home or away. These are important to add because the player stats require both week number and home/away teams to access, which will kinda suck to do. I'm working on it and will try to figure out how to add players by tomorrow evening.