Create login.js
This file provides some functionality for someone to create a profile, which then automatically logs them in if successfully created, then when they are logged in (after creating a profile). Notice they are given an option while logging in to create a league. If they check yes and the create account is successful, then they are redirected to creating a league. One other note, in order for this to appear, the HTML file must have the main div id be named "login".